+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.15:
 
 * #17 Upgrade to bootique-jetty 0.14
+* #18 Remove API's deprecated since 0.11
 
 ## 0.14: